添加了之前Integer类型的构造方法，兼容之前版本的用法，同时修改了更多处的int类型变量为long类型，避免超出范围的问题

+++ TaskExecutor/src/main/java/com/github/mayp1998/taskExecutor/util/LimitedTaskRunnable.java
@@ -10,9 +10,9 @@ public class LimitedTaskRunnable implements Runnable {
     private Runnable runnable;
 
     private boolean type;
-    private int runNumber;
+    private long runNumber;
 
-    private int run;// 已运行次数
+    private long run;// 已运行次数
 
     public LimitedTaskRunnable(Task task) {
         this.task = task;
